Added new features to game that include a player freeze feature and a new game piece added to game board. Also created new images for player that change depending on what the user has selected. Modified both backend and frontend code to add these features. Also added comments to a few files.

+++ SnowbotRush/src/main/java/com/game/util/FileReaderWriter.java
@@ -3,48 +3,47 @@ package com.game.util;
 import java.io.*;
 
 public class FileReaderWriter {
+    // method to help save the file to server
     static public void saveNewFile(File dir, String name, String content) throws FileNotFoundException {
-        File myFile = null;
+        File myFile = null; // declare file handle
         try {
-            myFile = new File(dir, name);
-            myFile.delete();
-            myFile.createNewFile();
+            myFile = new File(dir, name); // get the file
+            if (myFile.isFile()) { // check this is a file
+                myFile.delete(); // delete the file
+                myFile.createNewFile(); // create a new file
+            }
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException(e); // throw exception if any error
         }
 
-        if (myFile!=null){
-            PrintWriter output = new PrintWriter(myFile);
-            output.print(content);
-            output.close();
+        if (myFile!=null){ // check for null
+            PrintWriter output = new PrintWriter(myFile); // create the print write object to write the content to the file
+            output.print(content); // send the content to the file on the server
+            output.close(); // close the print writer
         }
     }
 
-    /**
-     * Method reads the content of the file and returns it as String
-     * ***/
+    // method reads the content of the file and returns it as string
     static public String readHighScoresFile(File dir, String name) throws FileNotFoundException {
-        File myFile = null;
+        File myFile = null; // declare file handle
         try {
-            myFile = new File(dir, name);
-            if (myFile.createNewFile()) {
-                return null;
+            myFile = new File(dir, name); // get the file
+            if (myFile.createNewFile()) { // if a new file, return empty string
+                return "";
             } else {
-                FileReader fileInput = new FileReader(myFile);
-                BufferedReader input = new BufferedReader(fileInput);
+                FileReader fileInput = new FileReader(myFile); // create file reader
+                BufferedReader input = new BufferedReader(fileInput); // create buffered reader to read the file content
 
-                // read the content
-                StringBuffer buffer = new StringBuffer();
+                // read the content line by line from the file
+                StringBuffer buffer = new StringBuffer(); // create buffered string
                 String line;
-                while ((line = input.readLine()) != null) {
-                    buffer.append(line);
+                while ((line = input.readLine()) != null) { // continue reading until end of file is reached
+                    buffer.append(line); // store the line in buffer
                 }
-                String content = buffer.toString();
-
-                return content;
+                return buffer.toString(); // output the contents of the buffer as string
             }
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            throw new RuntimeException(e); // throw exception if any error
         }
     }
 }
